+++ src/oldnews/screens/new_subscription.py
@@ -110,7 +110,12 @@ class NewSubscription(ModalScreen[NewSubscriptionData | None]):
                 )
             )
             with Horizontal():
-                yield Button("Add", id="add", variant="primary", disabled=True)
+                yield Button(
+                    add_key("Add", "Enter", self),
+                    id="add",
+                    variant="primary",
+                    disabled=True,
+                )
                 yield Button(
                     add_key("Cancel", "Esc", self), id="cancel", variant="error"
                 )
@@ -125,9 +130,10 @@ class NewSubscription(ModalScreen[NewSubscriptionData | None]):
     @on(Input.Changed, "#feed")
     def _refresh_state(self) -> None:
         """Refresh the state of the dialog."""
-        self.add_button.disabled = not bool(self.feed_input.value.strip())
+        self.add_button.disabled = not looks_webish(self.feed_input.value.strip())
 
     @on(Button.Pressed, "#add")
+    @on(Input.Submitted)
     def action_add(self) -> None:
         """React to the user pressing the add button."""
         if feed := self.feed_input.value.strip():