+++ ChangeLog.md
@@ -14,6 +14,7 @@
   ([#135](https://github.com/davep/oldnews/pull/135))
 - Added the `UserInformation` command for showing all known information
   about the currently-logged-in user.
+  (#136[](https://github.com/davep/oldnews/pull/136))
 
 ## v0.10.0