+++ ChangeLog.md
@@ -26,6 +26,7 @@
   ([#140](https://github.com/davep/oldnews/pull/140))
 - Improved the dialog that gets the name of a folder when moving a
   subscription. ([#156](https://github.com/davep/oldnews/pull/156))
+- Tidied up the captions of various dialog buttons.
 
 ## v0.10.0